created proper file tree, but broke the app in the process. unsure of where
